refactor: separate TurnResult
