+++ src/main/kotlin/omok/model/domain/gameState/WhiteTurn.kt
@@ -1,7 +1,5 @@
 package omok.model.domain.gameState
 
-import omok.model.domain.rule.RuleMapper
-import omok.model.domain.rule.WhiteWinRule
 import omok.model.entity.Stone
 import omok.model.entity.position.Position
 
@@ -12,19 +10,7 @@ object WhiteTurn : Turn {
         gameState: GameState,
         position: Position,
     ): GameState {
-        val adaptedBoard = RuleMapper.adapt(gameState.board)
-        val adaptedPosition = RuleMapper.adapt(position)
-        gameState.board.put(position, stone)
-        return nextGameState(adaptedBoard, adaptedPosition, gameState)
+        gameState.board.put(position, Stone.WHITE)
+        return TurnResult(gameState, position).whiteTurnResult()
     }
-
-    private fun nextGameState(
-        adaptedBoard: List<List<Int>>,
-        adaptedPosition: Pair<Int, Int>,
-        gameState: GameState,
-    ): GameState =
-        when {
-            WhiteWinRule.validated(adaptedBoard, adaptedPosition) -> gameState.copy(playing = false)
-            else -> gameState.copy(turn = BlackTurn())
-        }
 }
